LayoutTests/platform/glib/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/operator-dictionary-movablelimits-006-expected.txt


PASS Operator dictionary chunk 1 - movablelimits
PASS Operator dictionary chunk 2 - movablelimits
FAIL Operator dictionary chunk 3 - movablelimits assert_approx_equals: Movablelimits property for ⨞ prefix should be 'true' expected 50 +/- 1 but got 60
PASS Operator dictionary chunk 4 - movablelimits
movablelimits for "⨞" (prefix):
⨞
 
 VS
⨞
 
